requirements.txt
tensorflow==2.16.1
numpy==1.26.4
opencv-python==4.9.0.80
keras==3.1.1
speechrecognition==3.10.3
pyaudio==0.2.14
pillow==10.3.0
pyttsx3==2.90